pyproject.toml
[project]
name = "massGui"
version = "0.0.1"
authors = [
  { name="Grant Mondeel", email="gmondee@g.clemson.edu" },
]
description = "Mass GUI for Boulder 2023"
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
  "numpy",
  "PyQt5",
  "pytest-qt",
  "logging",
  "matplotlib"
]

[project.scripts]
massGui = "massGui:main"
massGuiTest = "massGui:runOpen"

[project.entry-points]


[project.urls]
"Homepage" = "https://github.com/gmondee/massGui"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = ["-ra", "--showlocals", "--strict-markers", "--strict-config", "--import-mode=importlib"]
testpaths = [
    "tests"]
log_cli = 1